Module Ast__Ast_types

These are the types used in constructing any AST for a Bolt program. Bolt programs consist of a list of class definitions, followed by a list of trait definitions and finally an expression to execute.

type loc = Stdlib.Lexing.position

Stores the line and position of the token

module type ID = sig ... end

An abstract type for identifiers

module Var_name : ID

module Class_name : ID

module Region_name : ID

module Field_name : ID

module Method_name : ID

module Function_name : ID

type capability =

| Linear
Only permitted one alias to the object at any time

| Thread
Permitted multiple aliases but only within the same thread

| Read
Allowed access through multiple aliases but the object must be immutable

| Subordinate
Only accessible from within encapsulating object - inherits capability of owner

| Locked
Freely sharable amongst threads, operations protected by a lock on the object

Define capabilities for data references

type mode =

| MConst
Immutable

| MVar
Mutable

Determines whether field is (im)mutable

type type_expr =

| TEInt

| TEClass of Class_name.t

| TEVoid

| TEBool

Define types of expressions in Bolt programs

type field_defn =

| TField of mode * type_expr * Field_name.t * Region_name.t list

Class Field declarations are of the form "mode name : type" e.g. const f : int

type region =

| TRegion of capability * Region_name.t

Regions consist of name and the capability

type param =

| TParam of type_expr * Var_name.t * Region_name.t list option

| TVoid

Parameter of a function optionally has a region guard

val string_of_loc : loc -> string

val string_of_cap : capability -> string

val string_of_mode : mode -> string

val string_of_type : type_expr -> string

type bin_op =

| BinOpPlus

| BinOpMinus

| BinOpMult

| BinOpIntDiv

| BinOpRem

| BinOpLessThan

| BinOpLessThanEq

| BinOpGreaterThan

| BinOpGreaterThanEq

| BinOpAnd

| BinOpOr

| BinOpEq

| BinOpNotEq

val string_of_bin_op : bin_op -> string

type un_op =

| UnOpNot

| UnOpNeg

val string_of_un_op : un_op -> string