Module Ast__Pprint_ast

This module defines a set of pretty printing functions for the parts of the AST that are common across the parsed and typed ASTs - to be used by pretty printing functions of the two ASTs.

They all take in a formatter which specifies output channel e.g. a string formatter or a stdout formatter. (Think of this as a generalisation of printf)

The second argument for all of these is an indent - this corresponds to nesting depth within the AST

val pprint_capabilities : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.capability list -> unit

val pprint_capability_names : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.Capability_name.t list -> unit

val pprint_modifier : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.modifier -> unit

val pprint_field_defn : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.field_defn -> unit

val pprint_type_expr : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.type_expr -> unit

val pprint_params : Stdlib.Format.formatter -> indent:string -> Ast.Ast_types.param list -> unit